requirements.txt
alembic==1.13.2
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.10.2
apache-airflow-providers-amazon==8.28.0
apache-airflow-providers-apache-spark==4.10.0
apache-airflow-providers-celery==2.1.4
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.16.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-redis==2.0.4
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==2.1.3
apispec==6.6.1
argcomplete==2.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
Babel==2.10.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
blinker==1.8.2
boto3==1.35.24
botocore==1.35.24
cachelib==0.9.0
cattrs==1.10.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==6.8.2
comm==0.2.2
commonmark==0.9.1
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.0
cron-descriptor==1.2.24
croniter==3.0.3
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==1.0.0
fonttools==4.53.1
fsspec==2024.9.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.1.0
idna==3.3
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
kombu==5.2.4
lazy-object-proxy==1.7.1
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.0
Mako==1.2.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
nest-asyncio==1.6.0
numpy==2.0.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
pluggy==1.5.0
ply==3.11
prison==0.2.1
prometheus-client==0.14.1
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
PyAthena==3.9.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.5.2
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==26.2.0
redis==3.5.3
redshift-connector==2.1.3
referencing==0.35.1
requests==2.27.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==12.4.4
rich-argparse==1.5.2
rpds-py==0.20.0
s3transfer==0.10.2
scramp==1.4.5
setproctitle==1.3.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.38.2
sqlparse==0.5.1
stack-data==0.6.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.2.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==1.26.9
vine==5.0.0
watchtower==3.3.1
wcwidth==0.2.5
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
hadoop-aws
aws-java-sdk